+++ CMakeLists.txt
@@ -2821,7 +2821,6 @@ foreach(_hdr
   include/grpcpp/security/credentials.h
   include/grpcpp/security/credentials_impl.h
   include/grpcpp/security/server_credentials.h
-  include/grpcpp/security/server_credentials_impl.h
   include/grpcpp/security/tls_credentials_options.h
   include/grpcpp/server.h
   include/grpcpp/server_builder.h
@@ -3515,7 +3514,6 @@ foreach(_hdr
   include/grpcpp/security/credentials.h
   include/grpcpp/security/credentials_impl.h
   include/grpcpp/security/server_credentials.h
-  include/grpcpp/security/server_credentials_impl.h
   include/grpcpp/security/tls_credentials_options.h
   include/grpcpp/server.h
   include/grpcpp/server_builder.h
