+++ Makefile
@@ -4971,7 +4971,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/security/credentials.h \
     include/grpcpp/security/credentials_impl.h \
     include/grpcpp/security/server_credentials.h \
-    include/grpcpp/security/server_credentials_impl.h \
     include/grpcpp/security/tls_credentials_options.h \
     include/grpcpp/server.h \
     include/grpcpp/server_builder.h \
@@ -5663,7 +5662,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/security/credentials.h \
     include/grpcpp/security/credentials_impl.h \
     include/grpcpp/security/server_credentials.h \
-    include/grpcpp/security/server_credentials_impl.h \
     include/grpcpp/security/tls_credentials_options.h \
     include/grpcpp/server.h \
     include/grpcpp/server_builder.h \
