+++ build_autogenerated.yaml
@@ -2154,7 +2154,6 @@ libs:
   - include/grpcpp/security/credentials.h
   - include/grpcpp/security/credentials_impl.h
   - include/grpcpp/security/server_credentials.h
-  - include/grpcpp/security/server_credentials_impl.h
   - include/grpcpp/security/tls_credentials_options.h
   - include/grpcpp/server.h
   - include/grpcpp/server_builder.h
@@ -2540,7 +2539,6 @@ libs:
   - include/grpcpp/security/credentials.h
   - include/grpcpp/security/credentials_impl.h
   - include/grpcpp/security/server_credentials.h
-  - include/grpcpp/security/server_credentials_impl.h
   - include/grpcpp/security/tls_credentials_options.h
   - include/grpcpp/server.h
   - include/grpcpp/server_builder.h
