+++ gRPC-C++.podspec
@@ -166,7 +166,6 @@ Pod::Spec.new do |s|
                       'include/grpcpp/security/credentials.h',
                       'include/grpcpp/security/credentials_impl.h',
                       'include/grpcpp/security/server_credentials.h',
-                      'include/grpcpp/security/server_credentials_impl.h',
                       'include/grpcpp/security/tls_credentials_options.h',
                       'include/grpcpp/server.h',
                       'include/grpcpp/server_builder.h',
