+++ src/cpp/server/external_connection_acceptor_impl.h
@@ -23,7 +23,6 @@
 
 #include <grpc/impl/codegen/grpc_types.h>
 #include <grpcpp/security/server_credentials.h>
-#include <grpcpp/security/server_credentials_impl.h>
 #include <grpcpp/server_builder.h>
 #include <grpcpp/support/channel_arguments.h>
