+++ tools/doxygen/Doxyfile.c++.internal
@@ -1026,7 +1026,6 @@ include/grpcpp/security/auth_metadata_processor.h \
 include/grpcpp/security/credentials.h \
 include/grpcpp/security/credentials_impl.h \
 include/grpcpp/security/server_credentials.h \
-include/grpcpp/security/server_credentials_impl.h \
 include/grpcpp/security/tls_credentials_options.h \
 include/grpcpp/server.h \
 include/grpcpp/server_builder.h \
